Min and max ranges computes correctly.
Ranges are not just split in two anymore. Also added some more debugging prints for PDU handler function

+++ src/headers/hash_table.h
@@ -13,7 +13,7 @@
  * kuba@cs.umu.se
  */
 
-typedef uint8_t hash_t;
+typedef uint64_t hash_t;
 
 struct table_entry {
     char* ssn;
